refactor(layout): implement glass morphism and unified tab styling
- Add TabSwitcher component matching LayoutSwitcher style
- Swap LeftRail and RightRail components to match visual positions
- Update LayoutSwitcher to display tabs horizontally
- Add GlassSurface component with liquid glass effect
- Update navigation items with SVG icons (Social, Trends, DeFi)
- Remove dropdown menus from navigation
- Apply consistent glass morphism styling across dashboard cards
- Update feed items to use GlassSurface component
