+++ src/components/layout/app-header/AppHeader.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
 import WebAppHeader from './WebAppHeader';
 import MobileAppHeader from './MobileAppHeader';
+import { useLayoutVariant } from '../../../contexts/LayoutVariantContext';
 
 export default function AppHeader() {
+  const { variant } = useLayoutVariant();
+
   return (
     <>
-      {/* Hide WebAppHeader - navigation is now in the left rail */}
-      {/* <WebAppHeader /> */}
+      {/* Show WebAppHeader only in minimal mode where side navigation is hidden */}
+      {variant === 'minimal' && <WebAppHeader />}
       <MobileAppHeader />
     </>
   );
